+++ libmonit/src/util/StringBuffer.c
@@ -155,7 +155,7 @@ int StringBuffer_replace(T S, const char *a, const char *b) {
                         size_t diff = bl - strlen(a);
                         if (diff > 0) {
                                 size_t required = (diff * n) + S->used + 1;
-                                if (required >= S->length) {
+                                if (required >= (size_t)S->length) {
                                         S->length = (int)required;
                                         RESIZE(S->buffer, S->length);
                                 }
